requirements.txt
fastapi==0.128.0
python-dotenv==1.2.1
pydantic==2.12.5
aiohttp==3.13.3
python-socketio==5.16.0
fsrs==6.3.0